rename motor processes to something more generic
