+++ vivarium/compartments/chemotaxis_master.py
@@ -26,7 +26,7 @@ from vivarium.processes.translation import Translation
 from vivarium.processes.degradation import RnaDegradation
 from vivarium.processes.complexation import Complexation
 from vivarium.processes.chemoreceptor_cluster import ReceptorCluster
-from vivarium.processes.Mears2014_flagella_activity import FlagellaActivity
+from vivarium.processes.flagella_activity import FlagellaActivity
 from vivarium.processes.membrane_potential import MembranePotential
 from vivarium.processes.division_volume import DivisionVolume
