+++ vivarium/plots/coarse_motor.py
