+++ vivarium/plots/flagella_activity.py
